requirements.txt
redis==3.5.3
dj_database_url==0.5.0
Django==3.0.9
django-mptt==0.11.0
django-redis==4.11.0
Pillow==7.1.2
whitenoise==5.0.1
wheel==0.34.2
psycopg2-binary==2.8.5